pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "handsomenet"
version = "0.1.0"
description = "HandsomeNet: lightweight hand pose estimation with joint tokens and graph attention."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Georgi Pastrakov" }
]
dependencies = [
  "numpy>=2.0",
  "opencv-python>=4.10",
  "pillow>=11.0",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "torch>=2.5",
  "torchvision>=0.20",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.13",
  "pytest>=8.3",
  "pytest-cov>=6.0",
  "ruff>=0.7",
]

[tool.hatch.build.targets.wheel]
packages = ["src/handsomenet"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"
